Refactor and update various components
- Reordered import statements in outbound.go to include simpleobfs transport.
- Cleaned up commented log statements in bruteforce.go and udp.go for consistency.
- Simplified conditional statements in optimizer.go and other files for better readability.
- Updated test cases in conn_sniffer_relay_test.go and dns_control_cache_cleanup_test.go for improved clarity and performance.
- Enhanced comments in config.go and other files to follow a consistent style.
- Added a new configuration file for golangci-lint to enforce coding standards.
- Introduced unit tests for section.go to ensure correct functionality of NewSectionItem and its string representation.
- Made minor adjustments to various files for better code quality and maintainability.

+++ component/dns/request_routing.go
@@ -98,7 +98,7 @@ func (b *RequestMatcherBuilder) addQType(f *config_parser.Function, values []uin
 		}
 		b.rules = append(b.rules, requestMatchSet{
 			Type:     consts.MatchType_QType,
-			Value:    uint16(value),
+			Value:    value,
 			Not:      f.Not,
 			Upstream: uint8(upstreamId),
 		})
